used dynamic import (@) to import files, restructure the app folder and used webp images instead of jpg
