+++ src/app/pages/Home/Features.tsx
@@ -1,0 +1,7 @@
+const Features = () => {
+  return (
+    <div>Features</div>
+  )
+}
+
+export default Features
